fix(canvas): getImageData/putImageData 完整有符号语义 + ImageData 对象面 (R34xx)
WPT pixel-manipulation worker 变体 42 → 55 pass：
- getImageData 有符号语义：负 dims 矩形反向（top-to-bottom left-to-right）、负/越界
  sx/sy 区域画布外透明、dst 偏移按源列−x；EnforceRange long 校验（0xffffffff → TypeError，
  修复 8.6GB OOM 崩溃）；零尺寸 IndexSizeError
- putImageData 7 参 dirty 矩形（负 dims 源/目标反向 + 越界裁剪）+ 非有限 TypeError；
  host 负坐标 i32 + 画布裁剪
- ImageData 对象：width/height/data 只读 defineProperty + colorSpace/pixelFormat 字段
- createImageData 零尺寸 IndexSizeError
- 剩余 2 = float16 color space（深缺口，记 master.md）

+++ crates/canvas/src/context/offscreen.rs
@@ -72,7 +72,7 @@ impl OffscreenCanvas {
     /// 返回 ImageData 作为 ImageBitmap 的像素载体（ImageBitmap Rust 类型尚未建立）。
     pub fn transfer_to_image_bitmap(&mut self) -> ImageData {
         let (w, h) = (self.context.width(), self.context.height());
-        let bitmap = self.context.get_image_data(0, 0, w, h);
+        let bitmap = self.context.get_image_data(0, 0, w as i32, h as i32);
         // spec：transfer 后 bitmap 清空（绘制状态保留——仅 bitmap 像素清零）。
         self.context.clear_bitmap();
         bitmap
